backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
celery[redis]==5.4.0
redis==5.1.1
boto3==1.35.19
aiohttp==3.10.5
httpx==0.27.2
websockets==13.0
python-dotenv==1.0.1
minio==7.2.9
Pillow==10.4.0
openai==1.47.0
qiniu==7.14.0
